pyproject.toml
[project]
name = "cloud-workstation"
version = "0.1.1"
requires-python = ">=3.8"
readme = "README.md"
description = ""
authors = [
    { name = "Damien Ramunno-Johnson", email = "damien@squareup.com" }  # Use this object format for authors
]


dependencies = [
    "google-cloud-workstations>=0.5.6,<1.0.0",
    "click>=8.1.7,<9.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "wheel>=0.43.0,<0.44.0",
    "rich>=13.7.1,<14.0.0",
    "grpcio==1.64.1",  # Pin grpcio to version 1.64.1 as 1.65.0+ have issues https://github.com/grpc/grpc/issues/37178
    "google-cloud-logging>=3.10.0,<4.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0,<9.0.0",
    "mypy>=1.10.0,<2.0.0",
    "mkdocstrings-python>=1.10.0,<2.0.0",
    "mkdocs-section-index>=0.3.9,<0.4.0",
    "mkdocs>=1.6.0,<2.0.0",
    "mkdocs-material>=9.5.20,<10.0.0",
    "markdown-include>=0.8.1,<0.9.0",
    "mkdocs-gen-files>=0.5.0,<0.6.0",
    "mkdocs-literate-nav>=0.6.1,<0.7.0",
    "pymdown-extensions>=10.8.1,<11.0.0",
    "mkdocs-include-markdown-plugin>=6.0.5,<7.0.0",
    "mkdocs-redirects>=1.2.1,<2.0.0",
    "mkdocs-click>=0.8.1,<0.9.0",
    "mkdocs-autorefs>=1.0.1,<2.0.0",
    "pytest-xdist>=3.6.1,<4.0.0",
    "types-pyyaml>=6.0.12.20240311,<7.0.0",
    "pytest-cov",
    "pytest-mock",
]

[project.scripts]
workstation = "workstation.cli:cli"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["I", "D"]
ignore = ["D104", "D100"]

[tool.pytest.ini_options]
addopts = "-n 4"
markers = ["integration: mark a test as an integration test."]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff.lint.isort]
force-sort-within-sections = false

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D103"]


[tool.coverage.run]
omit = [
    "src/workstation/cli/__main__.py"
]
